+++ apps/client/src/app/components/home-market/home-market.module.ts
@@ -1,6 +1,6 @@
 import { GfFearAndGreedIndexModule } from '@ghostfolio/client/components/fear-and-greed-index/fear-and-greed-index.module';
-import { GfBenchmarkModule } from '@ghostfolio/ui/benchmark/benchmark.module';
-import { GfLineChartModule } from '@ghostfolio/ui/line-chart/line-chart.module';
+import { GfBenchmarkComponent } from '@ghostfolio/ui/benchmark';
+import { GfLineChartComponent } from '@ghostfolio/ui/line-chart';
 
 import { CommonModule } from '@angular/common';
 import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
@@ -13,9 +13,9 @@ import { HomeMarketComponent } from './home-market.component';
   exports: [HomeMarketComponent],
   imports: [
     CommonModule,
-    GfBenchmarkModule,
+    GfBenchmarkComponent,
     GfFearAndGreedIndexModule,
-    GfLineChartModule,
+    GfLineChartComponent,
     NgxSkeletonLoaderModule
   ],
   schemas: [CUSTOM_ELEMENTS_SCHEMA]