+++ apps/client/src/app/components/position/position.module.ts
@@ -1,6 +1,6 @@
 import { GfSymbolModule } from '@ghostfolio/client/pipes/symbol/symbol.module';
-import { GfTrendIndicatorModule } from '@ghostfolio/ui/trend-indicator';
-import { GfValueModule } from '@ghostfolio/ui/value';
+import { GfTrendIndicatorComponent } from '@ghostfolio/ui/trend-indicator';
+import { GfValueComponent } from '@ghostfolio/ui/value';
 
 import { CommonModule } from '@angular/common';
 import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
@@ -18,8 +18,8 @@ import { PositionComponent } from './position.component';
     CommonModule,
     GfPositionDetailDialogModule,
     GfSymbolModule,
-    GfTrendIndicatorModule,
-    GfValueModule,
+    GfTrendIndicatorComponent,
+    GfValueComponent,
     MatDialogModule,
     NgxSkeletonLoaderModule,
     RouterModule