+++ apps/client/src/app/pages/register/register-page.module.ts
@@ -1,4 +1,4 @@
-import { GfLogoModule } from '@ghostfolio/ui/logo';
+import { GfLogoComponent } from '@ghostfolio/ui/logo';
 
 import { CommonModule } from '@angular/common';
 import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
@@ -13,7 +13,7 @@ import { ShowAccessTokenDialogModule } from './show-access-token-dialog/show-acc
   declarations: [RegisterPageComponent],
   imports: [
     CommonModule,
-    GfLogoModule,
+    GfLogoComponent,
     MatButtonModule,
     RegisterPageRoutingModule,
     RouterModule,