+++ libs/ui/src/lib/activity-type/activity-type.component.ts
@@ -1,6 +1,8 @@
 import { translate } from '@ghostfolio/ui/i18n';
 
+import { CommonModule } from '@angular/common';
 import {
+  CUSTOM_ELEMENTS_SCHEMA,
   ChangeDetectionStrategy,
   Component,
   Input,
@@ -10,11 +12,14 @@ import { Type as ActivityType } from '@prisma/client';
 
 @Component({
   changeDetection: ChangeDetectionStrategy.OnPush,
+  imports: [CommonModule],
+  schemas: [CUSTOM_ELEMENTS_SCHEMA],
   selector: 'gf-activity-type',
+  standalone: true,
   styleUrls: ['./activity-type.component.scss'],
   templateUrl: './activity-type.component.html'
 })
-export class ActivityTypeComponent implements OnChanges {
+export class GfActivityTypeComponent implements OnChanges {
   @Input() activityType: ActivityType;
 
   public activityTypeLabel: string;