+++ libs/ui/src/lib/trend-indicator/index.ts
@@ -1,1 +1,1 @@
-export * from './trend-indicator.module';
+export * from './trend-indicator.component';